refactor: separate note vs tags DB functions
